Fix benchmark of chain
